+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.1.1
+
+- Added a `fallback` language option
+- Added `URL::language` to get an URL to a language
+- Added `URL::switchLanguage` to get an URL to the current page in another language
+
 ## 1.1.0
 
 - `Lang::get` now loads translations in your default locale if none found in the current one